#172: 상품 세부사항 관련 Cell 각각 파일 분리
